requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0